Updated routes and moved app.js
