+++ app/index.js
@@ -4,7 +4,7 @@ import { Provider } from 'react-redux';
 import { Router, hashHistory } from 'react-router';
 import { syncHistoryWithStore } from 'react-router-redux';
 
-import routes from './routes.js';
+import routes from './routes';
 import configureStore from './store/configureStore';
 import './app.global.css';
