+++ app/routes.js
@@ -1,15 +1,15 @@
 import React from 'react';
 import { Route, IndexRoute } from 'react-router';
 
-import App from './containers/App';
-import HomePage from './containers/HomePage';
+import App from './App';
+import Dashboard from './routes/Dashboard';
 
 const appRoute = (
   <Route path="/" component={App}>
 
     {/* Load the homepage page by default */}
-    <IndexRoute to="homepage" />
-    {HomePage}
+    <IndexRoute to="dashboard" />
+    {Dashboard}
 
   </Route>
 );
