--- app/routes/Home.js
@@ -1,15 +1,0 @@
-// @flow
-import React, { Component } from 'react';
-// import { Link } from 'react-router';
-import styles from './Home.css';
-
-
-export default class Home extends Component {
-  render() {
-    return (
-      <div>
-        <div className={styles.container} data-tid="container" />
-      </div>
-    );
-  }
-}
